Move adapters to adapters, add api and auth
