+++ app/adapters/auth.js
@@ -1,0 +1,6 @@
+import config from "../config/environment";
+import ApplicationAdapter from "./application";
+
+export default ApplicationAdapter.extend({
+  host: config.authBaseUri
+});
